remove comment in details view
The isDeleted in the API is handled slightly different than the rest of the checkbox properties, I implemented it like this to match how the other checkbox values are displayed. Type = text forces the input to show true or false as a value.

for example: The checkbox values on the milestone in the API are converted between nullable ints and used as nullable bools on the views, however isDeleted is a derived property from the base entity and it is always a non nullable bool.