Create temporary javascript logic for disable logic
We do not have an answer for "disable on check" for checkbox. The current functionality only works for "enable on check".  This code handles that functionality for section #39 in the A1a

In addition, I added logic to disallow users from submitting contradicting answers by submitting "not applicable" but also submitting other answers for the #39 section of form A1a