Applied routing error check
The OnGet visit index will now require a param of statuses or provide an error. This will help for development to prevent an incorrect page-route to the visit index. Without it, the visit index pagination will not work correctly and serves as making sure there is always a default statuses filter set.