Draft code: resolve issue with input & radio combinations for range fields
There are different combinations of parent types that control child types. A parent can be a input type or radio type and a child can be a input type or radio type, this differentiation is important because we need to get the value of the parent and child elements. So we need to specifically get ":checked" for a radio type or we won't have a value to validate.

Originally, I had radio button logic and input logic sperate, but this can be merged in the addmethod and adapter methods