Required on range clear (#482)
This squash commit contains a major version update in client-side validation

* Required on range clear

* Use validate method to clear styling and manually clear error messages for watched children

* Address LINT errors

* Update method of finding target elements for error message

* Adjust naming for targets in required if range adapter

* Remove console log log statements from custom validation

* Merge main and remove extra call to custom validation in C2 partial

When removing the extra call to custom validation.js stopped the validation message from rendering in a way where the jquery did not react to it

Validation errors with a single span error message did not trigger validation.js. Without the extra call error validation messages now render as nested span elements and react as expected and in turn removing error validation messagse with my code

* Adjust dev comment wording

* Address LINT errors

* Address LINT error

* Remove class add and make required if range exclude high and low ints as out of range

* create custom unobtrusive file, remove unnecessary code, and remove validation summary lines by input

* Pushing up code to prepare for pairing + examples

The code added will help describe what is happening, pushing up temporary "debug" code so we can pair on the issue

* Add validation attributes to table cell radio buttons tag helper

* Apply update to radio tag helpers to generate asp-for attributes from unobtrusive

* Adjust stylings for valid and invalid radio buttons

* Adjust bugs in JavaScript custom validation

* Draft code: resolve issue with input & radio combinations for range fields

There are different combinations of parent types that control child types. A parent can be a input type or radio type and a child can be a input type or radio type, this differentiation is important because we need to get the value of the parent and child elements. So we need to specifically get ":checked" for a radio type or we won't have a value to validate.

Originally, I had radio button logic and input logic sperate, but this can be merged in the addmethod and adapter methods

* Refactor and clean up code in custom validation

* Create requiredifregex attribute and begin adding to property

* Implement requiredifregex into A5D2

* Clean up code for requiredifregex

* Add removal of validation summary messages and remove unobtrusive.custom.js

* Allow required if unobtrusive to remove error messages

* update some properties to use regex attribute and adjust custom validation

* Adjust property validation attributes and address LINT errors

* Address Javascript LINT errors

* Address javascript LINT errors

* Reformat code and add checks

Reformat code, add checkbox check for requiredif, remove error summary upon required field input

* Add semi colons and format code

* Add disable styling to radio buttons affected by unobtrusive validation

* Update adapters to use validation message remove function and update requiredif method / adapter

* Fix linter errors

* Remove validation messages for birthmonth, birthyear, and prohibited characters

* Fix tests to locate error elements

* Fills valid valid values for other form sections, so server-side MINT validation can be tested

* Add back in removed table row ending element

* Update to major version 14.0.0

---------

Co-authored-by: Ashley